+++ packages/analytics-js/__fixtures__/fixtures.ts
@@ -1,5 +1,6 @@
-import type { SourceConfigResponse } from '@rudderstack/analytics-js/components/configManager/types';
+import type { SourceConfigResponse } from '../src/components/configManager/types';
 import { COOKIE_KEYS } from '@rudderstack/analytics-js-cookies/constants/cookies';
+import type { StorageEntries } from '@rudderstack/analytics-js-common/types/ApplicationState';
 
 const identifyRequestPayload = {
   userId: '123456',
@@ -39,15 +40,6 @@ const screenRequestPayload = {
     screenSize: 10,
     title: 'Pizza',
     referrer: 'https://google.com',
-  },
-};
-
-const groupRequestPayload = {
-  userId: '12345',
-  groupId: '1',
-  traits: {
-    name: 'Company',
-    description: 'Google',
   },
 };
 
@@ -93,9 +85,6 @@ const dummySourceConfigResponse: SourceConfigResponse = {
     enabled: true,
     workspaceId: 'dummyWorkspaceId',
     updatedAt: '2023-02-03T13:53:35.731Z',
-    dataplanes: {
-      US: '',
-    },
     destinations: [
       {
         config: {
@@ -269,7 +258,7 @@ const entriesWithOnlyCookieStorage = {
     type: 'cookieStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithOnlyLocalStorage = {
   userId: {
@@ -308,7 +297,7 @@ const entriesWithOnlyLocalStorage = {
     type: 'localStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithoutCookieStorage = {
   userId: {
@@ -347,7 +336,7 @@ const entriesWithoutCookieStorage = {
     type: 'localStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithoutCookieAndLocalStorage = {
   userId: {
@@ -386,7 +375,7 @@ const entriesWithoutCookieAndLocalStorage = {
     type: 'sessionStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithOnlyNoStorage = {
   userId: {
@@ -425,7 +414,7 @@ const entriesWithOnlyNoStorage = {
     type: 'none',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithOnlySessionStorage = {
   userId: {
@@ -464,7 +453,7 @@ const entriesWithOnlySessionStorage = {
     type: 'sessionStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithMixStorage = {
   userId: {
@@ -503,7 +492,7 @@ const entriesWithMixStorage = {
     type: 'memoryStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithMixStorageButWithoutNone = {
   userId: {
@@ -542,7 +531,7 @@ const entriesWithMixStorageButWithoutNone = {
     type: 'memoryStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const loadOptionWithEntry = {
   userId: {
@@ -557,7 +546,7 @@ const loadOptionWithEntry = {
   sessionInfo: {
     type: 'none',
   },
-};
+} satisfies StorageEntries;
 
 const postConsentStorageEntryOptions = {
   userId: {
@@ -572,7 +561,7 @@ const postConsentStorageEntryOptions = {
   sessionInfo: {
     type: 'none',
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithInMemoryFallback = {
   userId: {
@@ -611,7 +600,7 @@ const entriesWithInMemoryFallback = {
     type: 'memoryStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const loadOptionWithInvalidEntry = {
   userId: {
@@ -623,7 +612,7 @@ const loadOptionWithInvalidEntry = {
   anonymousId: {
     type: 'cookieStorage',
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithStorageOnlyForSession = {
   userId: {
@@ -662,7 +651,7 @@ const entriesWithStorageOnlyForSession = {
     type: 'none',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const anonymousIdWithNoStorageEntries = {
   userId: {
@@ -701,7 +690,7 @@ const anonymousIdWithNoStorageEntries = {
     type: 'cookieStorage',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 const entriesWithStorageOnlyForAnonymousId = {
   userId: {
@@ -740,7 +729,7 @@ const entriesWithStorageOnlyForAnonymousId = {
     type: 'none',
     key: COOKIE_KEYS.authToken,
   },
-};
+} satisfies StorageEntries;
 
 export {
   identifyRequestPayload,