+++ packages/analytics-js/__tests__/components/userSessionManager/UserSessionManager.test.ts
@@ -67,7 +67,7 @@ describe('User session manager', () => {
 
   const setDataInCookieStorageEngine = (data: any) => {
     Object.entries(data).forEach(([key, value]) => {
-      clientDataStoreCookie.engine.setItem(key, value);
+      clientDataStoreCookie.engine.setItem(key, value as string);
     });
   };
 
@@ -113,6 +113,10 @@ describe('User session manager', () => {
       state.loadOptions.value.sessions = {
         autoTrack: true,
         timeout: 10000,
+        cutOff: {
+          enabled: false,
+          duration: 12 * 60 * 60 * 1000,
+        },
       };
       state.storage.entries.value = entriesWithStorageOnlyForAnonymousId;
       userSessionManager.init();
@@ -134,8 +138,7 @@ describe('User session manager', () => {
         autoTrack: true,
         expiresAt: expect.any(Number),
         id: expect.any(Number),
-        sessionStart: undefined,
-        timeout: 10000, // TODO: fix this after this entry is removed from state
+        timeout: 10000,
       });
       // This also covers the data migration from previous storage to current
       expect(state.session.anonymousId.value).toBe(customData.rl_anonymous_id);
@@ -152,11 +155,12 @@ describe('User session manager', () => {
       userSessionManager.syncStorageDataToState();
       expect(spy).toHaveBeenCalledWith('anonId');
     });
-    it('should set anonymousId with existing logic if external name is not string', () => {
+    it('should set anonymousId with existing logic if external anonymous ID cookie name is not string', () => {
       const customData = {
         rl_anonymous_id: 'dummy-anonymousId',
       };
       setDataInCookieStorage(customData);
+      // @ts-expect-error - need to test this case
       state.loadOptions.value.externalAnonymousIdCookieName = 12345;
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       const spy = jest.spyOn(userSessionManager, 'getExternalAnonymousIdByCookieName');
@@ -171,6 +175,7 @@ describe('User session manager', () => {
         rl_anonymous_id: 'dummy-anonymousId',
       };
       setDataInCookieStorage(customData);
+      // @ts-expect-error - need to test this case
       state.loadOptions.value.externalAnonymousIdCookieName = null;
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       const spy = jest.spyOn(userSessionManager, 'getExternalAnonymousIdByCookieName');
@@ -216,7 +221,7 @@ describe('User session manager', () => {
         },
       };
       setDataInCookieStorage(customData);
-      state.loadOptions.value.sessions.autoTrack = false;
+      state.loadOptions.value.sessions!.autoTrack = false;
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.syncStorageDataToState();
       expect(state.session.sessionInfo.value).toStrictEqual({});
@@ -232,16 +237,16 @@ describe('User session manager', () => {
         },
       };
       setDataInCookieStorage(customData);
-      state.loadOptions.value.sessions.autoTrack = false;
+      state.loadOptions.value.sessions!.autoTrack = false;
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.syncStorageDataToState();
       expect(state.session.sessionInfo.value).toStrictEqual({
         id: 1726655503445,
         expiresAt: expect.any(Number),
-        timeout: 60000,
         manualTrack: true,
         autoTrack: false,
         sessionStart: false,
+        timeout: undefined,
       });
     });
     it('should set sessionInfo if sessionInfo exists in storage with autoTrack enabled', () => {
@@ -257,7 +262,7 @@ describe('User session manager', () => {
       setDataInCookieStorage(customData);
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.syncStorageDataToState();
-      expect(state.session.sessionInfo.value).toStrictEqual({
+      expect(state.session.sessionInfo.value).toEqual({
         id: 1726655503445,
         expiresAt: expect.any(Number),
         timeout: 1800000,
@@ -527,9 +532,9 @@ describe('User session manager', () => {
       expect(state.session.initialReferringDomain.value).toBe(
         customData.rl_page_init_referring_domain,
       );
-      expect(state.session.sessionInfo.value).toStrictEqual({
+      expect(state.session.sessionInfo.value).toEqual({
         ...customData.rl_session,
-        timeout: expect.any(Number), // TODO: fix this after this entry is removed from state
+        timeout: expect.any(Number),
       });
       expect(state.session.authToken.value).toBe(customData.rl_auth_token);
     });
@@ -549,7 +554,7 @@ describe('User session manager', () => {
       state.session.sessionInfo.value = {
         autoTrack: true,
         manualTrack: false,
-        id: 'new-session-id',
+        id: 1234567890,
       };
       state.session.authToken.value = 'new-auth-token';
 
@@ -565,7 +570,7 @@ describe('User session manager', () => {
       expect(clientDataStoreCookie.get('rl_session')).toStrictEqual({
         autoTrack: true,
         manualTrack: false,
-        id: 'new-session-id',
+        id: 1234567890,
       });
       expect(clientDataStoreCookie.get('rl_auth_token')).toBe('new-auth-token');
     });
@@ -583,12 +588,11 @@ describe('User session manager', () => {
       );
       expect(state.session.initialReferrer.value).toBe('$direct'); // referrer is recomputed
       expect(state.session.initialReferringDomain.value).toBe('');
-      expect(state.session.sessionInfo.value).toStrictEqual({
+      expect(state.session.sessionInfo.value).toEqual({
         autoTrack: true,
         expiresAt: expect.any(Number),
         id: expect.any(Number),
-        sessionStart: undefined,
-        timeout: expect.any(Number), // TODO: fix this after this entry is removed from state
+        timeout: expect.any(Number),
       });
       expect(state.session.authToken.value).toBe(DEFAULT_USER_SESSION_VALUES.authToken);
     });
@@ -603,8 +607,49 @@ describe('User session manager', () => {
       expect(state.session.sessionInfo.value).toStrictEqual({});
     });
 
+    it('should reset the session info to default value if no session tracking is configured', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.loadOptions.value.sessions = {
+        autoTrack: false,
+      };
+      userSessionManager.init();
+
+      expect(state.session.sessionInfo.value).toStrictEqual({});
+    });
+
+    it('should reset the cut off expiry timestamp (retrieved from storage) if cut off is disabled in the configuration', () => {
+      const customData = {
+        rl_session: {
+          autoTrack: true,
+          manualTrack: false,
+          id: 1234567890,
+          timeout: 10000,
+          expiresAt: Date.now() + 1000,
+          cutOff: {
+            enabled: true,
+            duration: 12 * 60 * 60 * 1000,
+            expiresAt: Date.now() + 10000,
+          },
+        },
+      };
+      setDataInCookieStorage(customData);
+
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        cutOff: {
+          enabled: false,
+        },
+      };
+
+      userSessionManager.init();
+
+      expect(state.session.sessionInfo.value.cutOff).toBeUndefined();
+    });
+
     it('should log a warning and use default timeout if provided timeout is not in number format', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
+      // @ts-expect-error need to test this case
       state.loadOptions.value.sessions.timeout = '100000';
       userSessionManager.init();
       expect(defaultLogger.warn).toHaveBeenCalledWith(
@@ -615,7 +660,7 @@ describe('User session manager', () => {
 
     it('should log a warning and disable auto tracking if provided timeout is 0', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      state.loadOptions.value.sessions.timeout = 0;
+      state.loadOptions.value.sessions!.timeout = 0;
       userSessionManager.init();
       expect(defaultLogger.warn).toHaveBeenCalledWith(
         'UserSessionManager:: The session timeout value is 0, which disables the automatic session tracking feature. If you want to enable session tracking, please provide a positive integer value for the timeout.',
@@ -625,11 +670,145 @@ describe('User session manager', () => {
 
     it('should log a warning if provided timeout is less than 10 seconds', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      state.loadOptions.value.sessions.timeout = 5000; // provided timeout as 5 second
+      state.loadOptions.value.sessions!.timeout = 5000; // provided timeout as 5 second
       userSessionManager.init();
       expect(defaultLogger.warn).toHaveBeenCalledWith(
         `UserSessionManager:: The session timeout value 5000 ms is less than the recommended minimum of 10000 ms. Please consider increasing the timeout value to ensure optimal performance and reliability.`,
       );
+    });
+
+    it('should log a warning and use default value if the provided cut off duration is not a positive integer', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.loadOptions.value.sessions!.cutOff = {
+        enabled: true,
+        // @ts-expect-error - need to test this case
+        duration: '100000',
+      };
+
+      userSessionManager.init();
+
+      expect(defaultLogger.warn).toHaveBeenCalledWith(
+        'UserSessionManager:: The session cut off duration value "100000" is not a number. The default cut off duration of 43200000 ms will be used instead.',
+      );
+
+      expect(state.session.sessionInfo.value.cutOff).toStrictEqual({
+        enabled: true,
+        duration: 12 * 60 * 60 * 1000,
+        expiresAt: expect.any(Number),
+      });
+    });
+
+    it('should disable cut off and log a warning if the provided cut off duration is less than the session timeout', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        timeout: 60 * 60 * 1000,
+        cutOff: {
+          enabled: true,
+          duration: 10 * 60 * 1000, // less than the session timeout
+        },
+      };
+
+      userSessionManager.init();
+
+      expect(defaultLogger.warn).toHaveBeenCalledWith(
+        'UserSessionManager:: The session cut off duration value "600000" ms is less than the session timeout value "3600000" ms. The cut off functionality will be disabled.',
+      );
+
+      expect(state.session.sessionInfo.value.cutOff).toBeUndefined();
+    });
+
+    it('should reset the session and cut off expiry timestamp if it is exceeded', () => {
+      const currentCutOffExpiryTimestamp = 99; // expired cut off
+
+      const customData = {
+        rl_session: {
+          autoTrack: true,
+          manualTrack: false,
+          id: 98,
+          timeout: 10000,
+          expiresAt: 105, // still not expired
+          cutOff: {
+            enabled: true,
+            duration: 12 * 60 * 60 * 1000,
+            expiresAt: currentCutOffExpiryTimestamp,
+          },
+        },
+      };
+      setDataInCookieStorage(customData);
+
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        timeout: 60 * 60 * 1000,
+        cutOff: {
+          enabled: true,
+          duration: 12 * 60 * 60 * 1000,
+        },
+      };
+
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+
+      jest.useFakeTimers();
+      jest.setSystemTime(100);
+
+      userSessionManager.init();
+
+      expect(state.session.sessionInfo.value).toStrictEqual({
+        autoTrack: true,
+        id: Date.now(),
+        timeout: 3600000,
+        expiresAt: Date.now() + 3600000,
+        cutOff: {
+          enabled: true,
+          duration: 12 * 60 * 60 * 1000,
+          expiresAt: Date.now() + 12 * 60 * 60 * 1000,
+        },
+      });
+
+      jest.useRealTimers();
+    });
+
+    it('should not reset the cut off expiry timestamp if it is not exceeded', () => {
+      const currentCutOffExpiryTimestamp = 10000; // not expired cut off
+
+      const customData = {
+        rl_session: {
+          autoTrack: true,
+          manualTrack: false,
+          id: 1234567890,
+          timeout: 10000,
+          cutOff: {
+            enabled: true,
+            duration: 12 * 60 * 60 * 1000,
+            expiresAt: currentCutOffExpiryTimestamp,
+          },
+        },
+      };
+      setDataInCookieStorage(customData);
+
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        timeout: 60 * 60 * 1000,
+        cutOff: {
+          enabled: true,
+          duration: 12 * 60 * 60 * 1000,
+        },
+      };
+
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+
+      jest.useFakeTimers();
+      jest.setSystemTime(100);
+
+      userSessionManager.init();
+
+      expect(state.session.sessionInfo.value.cutOff).toStrictEqual({
+        enabled: true,
+        duration: 12 * 60 * 60 * 1000,
+        expiresAt: currentCutOffExpiryTimestamp,
+      });
+
+      jest.useRealTimers();
     });
   });
 
@@ -1084,6 +1263,7 @@ describe('User session manager', () => {
     it('should reset the value to default value if the value is not an object', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
+      // @ts-expect-error - need to test this case
       userSessionManager.setUserTraits('dummy-user-traits');
       expect(state.session.userTraits.value).toStrictEqual(DEFAULT_USER_SESSION_VALUES.userTraits);
     });
@@ -1159,6 +1339,7 @@ describe('User session manager', () => {
     it('should reset the value to default value if the value is not an object', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
+      // @ts-expect-error - need to test this case
       userSessionManager.setGroupTraits('dummy-group-traits');
       expect(state.session.groupTraits.value).toStrictEqual(
         DEFAULT_USER_SESSION_VALUES.groupTraits,
@@ -1256,6 +1437,7 @@ describe('User session manager', () => {
     it('should reset the value to default value if the value is not provided', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
+      // @ts-expect-error - need to test this case
       userSessionManager.setAuthToken();
       expect(state.session.authToken.value).toBe(DEFAULT_USER_SESSION_VALUES.authToken);
     });
@@ -1315,6 +1497,10 @@ describe('User session manager', () => {
         expiresAt: pastTimestamp,
         id: 1683613729115,
         sessionStart: false,
+        cutOff: {
+          enabled: false,
+          duration: 12 * 60 * 1000,
+        },
       };
       userSessionManager.refreshSession();
       expect(state.session.sessionInfo.value).toEqual({
@@ -1323,6 +1509,10 @@ describe('User session manager', () => {
         id: expect.any(Number),
         expiresAt: expect.any(Number),
         sessionStart: true,
+        cutOff: {
+          enabled: false,
+          duration: 12 * 60 * 1000,
+        },
       });
     });
 
@@ -1415,7 +1605,7 @@ describe('User session manager', () => {
       };
       const sessionId = userSessionManager.getSessionId();
       expect(typeof sessionId).toBe('number');
-      expect(sessionId.toString().length).toBeGreaterThan(0);
+      expect(sessionId!.toString().length).toBeGreaterThan(0);
     });
 
     it('should return null for expired session', () => {
@@ -1484,6 +1674,10 @@ describe('User session manager', () => {
         expiresAt: Date.now() - 1000,
         id: 1683613729115,
         sessionStart: false,
+        cutOff: {
+          enabled: false,
+          duration: 12 * 60 * 60 * 1000,
+        },
       };
       userSessionManager.startOrRenewAutoTracking(state.session.sessionInfo.value);
       expect(state.session.sessionInfo.value).toEqual({
@@ -1492,6 +1686,10 @@ describe('User session manager', () => {
         expiresAt: expect.any(Number),
         id: expect.any(Number),
         sessionStart: undefined,
+        cutOff: {
+          enabled: false,
+          duration: 12 * 60 * 60 * 1000,
+        },
       });
     });
 
@@ -1547,22 +1745,31 @@ describe('User session manager', () => {
 
   describe('reset', () => {
     it('should reset user session to the initial value except anonymousId', () => {
+      jest.useFakeTimers();
+      jest.setSystemTime(0);
+
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       userSessionManager.setAnonymousId(dummyAnonymousId);
       const sessionInfoBeforeReset = JSON.parse(JSON.stringify(state.session.sessionInfo.value));
+
+      jest.advanceTimersByTime(1000);
       userSessionManager.reset();
+
       expect(state.session.userId.value).toEqual('');
       expect(state.session.userTraits.value).toEqual({});
       expect(state.session.groupId.value).toEqual('');
       expect(state.session.groupTraits.value).toEqual({});
       expect(state.session.anonymousId.value).toEqual(dummyAnonymousId);
+
       // new session will be generated
       expect(state.session.sessionInfo.value.autoTrack).toBe(sessionInfoBeforeReset.autoTrack);
       expect(state.session.sessionInfo.value.timeout).toBe(sessionInfoBeforeReset.timeout);
       expect(state.session.sessionInfo.value.expiresAt).not.toBe(sessionInfoBeforeReset.expiresAt);
       expect(state.session.sessionInfo.value.id).not.toBe(sessionInfoBeforeReset.id);
       expect(state.session.sessionInfo.value.sessionStart).toBe(undefined);
+
+      jest.useRealTimers();
     });
 
     it('should clear the existing anonymousId and set a new anonymousId with first parameter set to true', () => {
@@ -1583,10 +1790,42 @@ describe('User session manager', () => {
 
     it('should not start a new session with second parameter set to true', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        timeout: 10 * 60 * 1000,
+        cutOff: {
+          enabled: false,
+        },
+      };
+
       userSessionManager.init();
       const sessionInfoBeforeReset = JSON.parse(JSON.stringify(state.session.sessionInfo.value));
       userSessionManager.reset(true, true);
       expect(state.session.sessionInfo.value).toEqual(sessionInfoBeforeReset);
+    });
+
+    it('should retain previous session timeout and cut off information when auto tracking is enabled', () => {
+      state.loadOptions.value.sessions = {
+        autoTrack: true,
+        timeout: 10 * 60 * 1000,
+        cutOff: { enabled: true, duration: 15 * 60 * 60 * 1000 },
+      };
+
+      userSessionManager.init();
+
+      userSessionManager.reset();
+
+      expect(state.session.sessionInfo.value).toEqual({
+        autoTrack: true,
+        timeout: 10 * 60 * 1000,
+        expiresAt: expect.any(Number),
+        id: expect.any(Number),
+        cutOff: {
+          enabled: true,
+          duration: 15 * 60 * 60 * 1000,
+          expiresAt: expect.any(Number),
+        },
+      });
     });
   });
 
@@ -1717,7 +1956,11 @@ describe('User session manager', () => {
     });
     describe('Network request to Data service is successful', () => {
       it('should validate cookies are set from the server side', done => {
-        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.source.value = {
+          workspaceId: 'sample_workspaceId',
+          id: 'sample_source_id',
+          name: 'sample_source_name',
+        };
         state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
         state.storage.cookie.value = {
           maxage: 10 * 60 * 1000, // 10 min
@@ -1755,7 +1998,11 @@ describe('User session manager', () => {
         }, 1000);
       });
       it('should set cookies from client side if not successfully set from the server side', done => {
-        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.source.value = {
+          workspaceId: 'sample_workspaceId',
+          id: 'sample_source_id',
+          name: 'sample_source_name',
+        };
         state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
         state.storage.cookie.value = {
           maxage: 10 * 60 * 1000, // 10 min
@@ -1783,7 +2030,11 @@ describe('User session manager', () => {
     });
 
     it('should set cookie from client side if data service is down', done => {
-      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.source.value = {
+        workspaceId: 'sample_workspaceId',
+        id: 'sample_source_id',
+        name: 'sample_source_name',
+      };
       state.serverCookies.dataServiceUrl.value =
         'https://dummy.dataplane.host.com/serverDown/rsaRequest';
       state.storage.cookie.value = {
@@ -1803,7 +2054,11 @@ describe('User session manager', () => {
       }, 1000);
     });
     it('should set cookie from client side if dataServerUrl is invalid', done => {
-      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.source.value = {
+        workspaceId: 'sample_workspaceId',
+        id: 'sample_source_id',
+        name: 'sample_source_name',
+      };
       state.serverCookies.dataServiceUrl.value =
         'https://dummy.dataplane.host.com/invalidUrl/rsaRequest';
       state.storage.cookie.value = {
@@ -1823,7 +2078,11 @@ describe('User session manager', () => {
       }, 1000);
     });
     it('should set cookie from client side if any unhandled error ocurred in serServerSideCookie function', () => {
-      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.source.value = {
+        workspaceId: 'sample_workspaceId',
+        id: 'sample_source_id',
+        name: 'sample_source_name',
+      };
       state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
       userSessionManager.getEncryptedCookieData = jest.fn(() => {
         throw new Error('test error');
@@ -1862,7 +2121,11 @@ describe('User session manager', () => {
     describe('makeRequestToSetCookie', () => {
       it('should make external request to exposed endpoint', done => {
         state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
-        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.source.value = {
+          workspaceId: 'sample_workspaceId',
+          id: 'sample_source_id',
+          name: 'sample_source_name',
+        };
         state.storage.cookie.value = {
           maxage: 10 * 60 * 1000, // 10 min
           path: '/',