+++ packages/analytics-js/src/components/core/Analytics.ts
@@ -14,7 +14,6 @@ import type { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
 import type { ApiObject } from '@rudderstack/analytics-js-common/types/ApiObject';
 import type {
   AnonymousIdOptions,
-  ConsentOptions,
   LoadOptions,
 } from '@rudderstack/analytics-js-common/types/LoadOptions';
 import type { ApiCallback } from '@rudderstack/analytics-js-common/types/EventApi';
@@ -69,6 +68,7 @@ import type { IAnalytics } from './IAnalytics';
 import { getConsentManagementData, getValidPostConsentOptions } from '../utilities/consent';
 import { dispatchSDKEvent, isDataPlaneUrlValid, isWriteKeyValid } from './utilities';
 import { safelyInvokeCallback } from '../utilities/callbacks';
+import type { ConsentOptions } from '@rudderstack/analytics-js-common/types/Consent';
 
 /*
  * Analytics class with lifecycle based on state ad user triggered events