+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.14.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.13.2...@rudderstack/analytics-js-integrations@3.14.0) (2025-05-09)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.19.0`
+
+### Features
+
+* onboarding userpilot integration ([#2103](https://github.com/rudderlabs/rudder-sdk-js/issues/2103)) ([9065069](https://github.com/rudderlabs/rudder-sdk-js/commit/90650693cd477ff3987174f8f6ec5972b034f2cb))
+
 ## [3.13.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.13.1...@rudderstack/analytics-js-integrations@3.13.2) (2025-04-25)
 
 ### Dependency Updates