+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.51.2](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.51.1...rudder-sdk-js@2.51.2) (2025-05-09)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.19.0`
 ## [2.51.1](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.51.0...rudder-sdk-js@2.51.1) (2025-04-25)
 
 ### Dependency Updates